% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beetGrowth}
\alias{beetGrowth}
\title{Growth of sugarbeet in weed-infested and weed-free conditions}
\format{
A data frame with 18 rows and 3 variables
}
\usage{
data("beetGrowth")
}
\description{
A data frame with 18 observations on the following 3 variables:
}
\details{
\itemize{
  \item DAE. numeric: Days After Emergence
  \item weightInf. numeric: weight of crop biomass on the weed infested plots
  \item weightFree. numeric: weight of crop biomass on the weed free plots
 }
}
\examples{
data(beetGrowth)
mod3 <- nls(weightInf ~ NLS.L3(DAE, b, c, d), data = beetGrowth)

}
\references{
Covarelli, G., Onofri, A., 1998. Effects of timing of weed removal and emergence in sugarbeet, in: Proceedings 6th EWRS Mediterranean Symposium, Montpellier, 13-15 May 1998. EWRS, pp. 65–72.
}
\keyword{datasets}
