% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{degradation}
\alias{degradation}
\title{Soil degradation kinetic for a herbicide}
\format{
A data frame with 24 rows and 2 variables
}
\usage{
data("degradation")
}
\description{
A data frame with 24 observations on the following 2 variables:
}
\details{
\itemize{
  \item Time. numeric: Days from start of incubation
  \item Conc. numeric: residual concentration
 }
}
\examples{
data("degradation")
degradation

}
\keyword{datasets}
