% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAgroData.R
\name{getAgroData}
\alias{getAgroData}
\title{Get one of the available datasets}
\usage{
getAgroData(fileName, where = "https://www.casaonofri.it/_datasets/",
                    type = "csv")
}
\arguments{
\item{fileName}{character: the name of the dataset (with no extension)}
\item{where}{character: the name of the repository, containing the fileName}
\item{type}{character: the extension of the fileName (with no dot)}
}
\value{
returns a data.frame
}
\description{
This function returns the dataset shown in the website or book and used to run
all the examples.
}
\examples{
getAgroData("rimsulfuron")

}
\author{
Andrea Onofri
}
