% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Get one of the available datasets}
\usage{
getData(dataName)
}
\arguments{
\item{dataName}{character: the name of the dataset (with no extension)}
}
\value{
returns a data.frame
}
\description{
This function returns the dataset shown in the website or book and used to run
all the examples.
}
\examples{
getData("rimsulfuron")

}
\author{
Andrea Onofri
}
