% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metamitron}
\alias{metamitron}
\title{Degradation of metamitron in soil with co-applied herbicides}
\format{
A data frame with 96 rows and 3 variables
}
\usage{
data("metamitron")
}
\description{
The dataset describes the degradation of metamitron in soil at 20°C
with several co-applied herbicides. It is a synthetic dataset,
that was generated by Monte Carlo methods, starting from the
observed data in Vischetti et al., 1996.
A data frame with 96 observations on the following 3 variables:
}
\details{
\itemize{
  \item Time. numeric: Days from start of incubation
  \item Herbicide. factor: M is 'metamitron', MP is 'metamitron+phenmedipham', MC is 'metamitron+chloridazon' and MPC is 'metamitron+phenmedipham+chloridazon'
  \item Conc. numeric: residual concentration of metamitron
 }
}
\examples{
data("metamitron")
metamitron

}
\references{
Vischetti, C., Marini, M., Businelli, M., Onofri, A., 1996. The effect of temperature and co-applied herbicides on the degradation rate of phenmedipham, chloridazon and metamitron in a clay loam soil in the laboratory, in: Re, A.D., Capri, E., Evans, S.P., Trevisan, M. (Eds.), “The Environmental Phate of Xenobiotics”, Proceedings X Symposium on Pesticide Chemistry, Piacenza. La Goliardica Pavese, Piacenza, pp. 287–294.
}
\keyword{datasets}
