% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mixture}
\alias{mixture}
\title{Efficacy of the mixture of two herbicides}
\format{
A data frame with 24 rows and 2 variables
}
\usage{
data("mixture")
}
\description{
Two herbicides are used alone and in mixture. It is a data frame
with 16 observations on the following 2 variables:
}
\details{
\itemize{
  \item Treat. Factor: the treatment levels
  \item Weight. numeric: the weight of weeds
 }
}
\examples{
data("degradation")
degradation

}
\keyword{datasets}
