% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAgroData.R
\name{getAgroData}
\alias{getAgroData}
\title{Get one of the available datasets}
\usage{
getAgroData(
  fileName,
  where = "https://www.casaonofri.it/_datasets/",
  type = "csv",
  fileEncoding = "UTF-8-BOM"
)
}
\arguments{
\item{fileName}{character: the name of the file (with no extension)}

\item{where}{character: the name of the web repository}

\item{type}{character: the extension of the web file}

\item{fileEncoding}{character: declares the encoding used on a file, deafults to "UTF-8-BOM"}
}
\value{
returns a data.frame
}
\description{
This function loads and returns a dataset available in an external repository
and stored as '.csv' or other type of text files.
all the examples.
}
\examples{
getAgroData("rimsulfuron")

}
\author{
Andrea Onofri
}
